requirements.txt
Django==1.7.1
django-jinja==1.0.5
Jinja2==2.7.3
Pillow==2.6.1
django-simple-captcha==0.4.3
django-ckeditor
